--- a/src/mono/wasm/runtime/wasm-simd-feature-detect.wat
+++ b/src/mono/wasm/runtime/wasm-simd-feature-detect.wat
@@ -1,0 +1,6 @@
+(module
+  (func $test (export "test")
+    (i32x4.splat (i32.const 0))
+    drop
+  )
+)