… src/mono/CMakeLists.txt
… src/mono/mono/mini/interp/interp-simd-intrins.def
… src/mono/mono/mini/interp/simd-methods.def
… src/mono/wasm/runtime/CMakeLists.txt
… src/mono/wasm/runtime/genmintops.py
… src/mono/wasm/runtime/jiterpreter-jit-call.ts
… src/mono/wasm/runtime/jiterpreter-opcodes.ts
… src/mono/wasm/runtime/jiterpreter-support.ts
… src/mono/wasm/runtime/jiterpreter-trace-generator.ts
… src/mono/wasm/runtime/jiterpreter.ts
… src/mono/wasm/runtime/wasm-simd-feature-detect.wasm
… src/mono/wasm/wasm.proj