@kg 2023-03-02
browser-bench timings with this rebased on main:

| measurement | jiterp (v128) | jiterp (no v128) | no interp simd |
|-:|-:|-:|-:|
|         Json, non-ASCII text serialize |     2.9236ms |     3.0078ms |     2.9899ms |
|       Json, non-ASCII text deserialize |     5.5887ms |     5.7569ms |     5.5366ms |
|                  Json, small serialize |     0.1248ms |     0.1266ms |     0.1287ms |
|                Json, small deserialize |     0.2142ms |     0.2190ms |     0.2100ms |
|                  Json, large serialize |    33.3725ms |    34.3400ms |    34.7770ms |
|                Json, large deserialize |    56.4239ms |    58.3258ms |    55.8387ms |
|                    Span, Reverse bytes |     0.0210ms |     0.0736ms |     0.0892ms |
|                    Span, Reverse chars |     0.1310ms |     0.1310ms |     0.0914ms |
|                    Span, IndexOf bytes |     0.7161us |     0.8906us |     0.8870us |
|                    Span, IndexOf chars |     0.1101ms |     0.1329ms |     0.0840ms |
|              Span, SequenceEqual bytes |     0.0325ms |     0.0248ms |     0.0889ms |
|              Span, SequenceEqual chars |     0.0647ms |     0.0496ms |     0.1781ms |
|                      String, Normalize |     1.4029ms |     1.3966ms |     1.4207ms |
|                String, Normalize ASCII |     0.1752ms |     0.1732ms |     0.1752ms |
|               Vector, Create Vector128 |     0.0942us |     0.0928us |     0.0902us |
|              Vector, Add 2 Vector128's |     0.1076us |     0.1062us |     0.1133us |
|         Vector, Multiply 2 Vector128's |     0.1076us |     0.1065us |     0.1133us |
|                Vector, Dot product int |     0.1004us |     0.0998us |     0.1069us |
|              Vector, Dot product ulong |     0.0891us |     0.0878us |     0.0891us |
|              Vector, Dot product float |     0.1117us |     0.1095us |     0.1109us |
|             Vector, Dot product double |     0.0916us |     0.0909us |     0.0916us |
|                      Vector, Sum sbyte |     0.1411us |     0.1407us |     0.1417us |
|                      Vector, Sum short |     0.1100us |     0.1109us |     0.1111us |
|                       Vector, Sum uint |     0.0907us |     0.0898us |     0.0912us |
|                     Vector, Sum double |     0.0898us |     0.0893us |     0.0900us |
|                      Vector, Min float |     0.1532us |     0.1549us |     0.1592us |
|                      Vector, Max float |     0.1673us |     0.1636us |     0.1646us |
|                     Vector, Min double |     0.1197us |     0.1180us |     0.1197us |
|                     Vector, Max double |     0.1209us |     0.1194us |     0.1237us |
|                Vector, Normalize float |     0.2620us |     0.2593us |     0.4157us |

My opinion is pretty mixed based on this. It looks like there are some workloads in our bench suite here that vectorize, and the performance improves (especially if we enable native use of wasm v128 opcodes). But a lot of stuff seems to hold steady or even be worse - presumably cases where the interpreter intrinsics support isn't broad enough, so saying 'v128 is hardware accelerated'! hurts us.

Not sure what the right approach to take is with this. Maybe if we had wider support for Vector128 at the interpreter intrinsic level we'd see more gains?

@BrzVlad what do you think? I think for now it makes sense to leave `INTERP_ENABLE_SIMD` guarded with `#if !HOST_BROWSER && __GNUC__`, and to land the jiterpreter support so that we have a starting point so we can consider enabling it later. The downside is that it could potentially bitrot in this state since it would never run.