@BrzVlad 2023-03-03
@kg I don't think the way you test is correct, since the numbers make no sense. It is like you are testing the same thing. What does `no interp simd` mean ? Are you enabling simd in csproj ? (`WasmEnableSIMD`, I think the linker trims out vectorized bcl code if not enabled)