@kg 2023-03-03
> @kg I don't think the way you test is correct, since the numbers make no sense. It is like you are testing the same thing. What does `no interp simd` mean ? Are you enabling simd in csproj ? (`WasmEnableSIMD`, I think the linker trims out vectorized bcl code if not enabled)

I enable interp SIMD for the first two columns by changing
`#if !HOST_BROWSER && __GNUC__`
to
`#if __GNUC__`
which does cause interpreter simd opcodes to appear in various places as expected.

I'll retest by editing the browser-bench csproj to add that property; because SIMD had been enabled on main I thought it would be sufficient to set the constant in the interp. If the code is being trimmed out that would explain not seeing much of a difference - the opcodes I'm seeing must be for stuff that escaped trimming.