@kg 2023-03-08
I still have no explanation for why this performs the way it does.

| measurement | time (this PR) | time (this+AOT) | time (main)
|-:|-:|-:|-:|
|         Json, non-ASCII text serialize |     3.8978ms |     0.2363ms |     4.1681ms |
|       Json, non-ASCII text deserialize |     6.2145ms |     0.4287ms |     6.3337ms |
|                  Json, small serialize |     0.1357ms |     0.0126ms |     0.1324ms |
|                Json, small deserialize |     0.2251ms |     0.0328ms |     0.2198ms |
|                  Json, large serialize |    36.5108ms |     3.5608ms |    35.9710ms |
|                Json, large deserialize |    59.1910ms |     9.3782ms |    58.2045ms |
|                    Span, Reverse bytes |     0.0219ms |     0.0021ms |     0.1244ms |
|                    Span, Reverse chars |     0.1289ms |     0.0053ms |     0.4433ms |
|                    Span, IndexOf bytes |     0.7610us |     0.1028us |     0.0010ms |
|                    Span, IndexOf chars |     0.1153ms |     0.0028ms |     0.0864ms |
|              Span, SequenceEqual bytes |     0.0400ms |     0.0022ms |     0.0295ms |
|              Span, SequenceEqual chars |     0.0797ms |     0.0042ms |     0.0590ms |
|                      String, Normalize |     1.4766ms |     0.7871ms |     1.4346ms |
|                String, Normalize ASCII |     0.1815ms |     0.0061ms |     0.1811ms |
|               Vector, Create Vector128 |     0.0975us |     0.0449us |     0.0930us |
|              Vector, Add 2 Vector128's |     0.1103us |     0.0479us |     0.1089us |
|         Vector, Multiply 2 Vector128's |     0.1117us |     0.0453us |     0.1100us |
|                Vector, Dot product int |     0.1060us |     0.0458us |     0.1036us |
|              Vector, Dot product ulong |     0.0933us |     0.0453us |     0.0910us |
|              Vector, Dot product float |     0.1162us |     0.0449us |     0.1152us |
|             Vector, Dot product double |     0.0971us |     0.0447us |     0.0940us |
|                      Vector, Sum sbyte |     0.1494us |     0.0449us |     0.1449us |
|                      Vector, Sum short |     0.1165us |     0.0447us |     0.1150us |
|                       Vector, Sum uint |     0.0961us |     0.0446us |     0.0945us |
|                     Vector, Sum double |     0.0947us |     0.0447us |     0.0922us |
|                      Vector, Min float |     0.2321us |     0.0451us |     0.2200us |
|                      Vector, Max float |     0.2233us |     0.0447us |     0.1786us |
|                     Vector, Min double |     0.1850us |     0.0451us |     0.1859us |
|                     Vector, Max double |     0.1447us |     0.0450us |     0.1349us |
|                Vector, Normalize float |     0.2761us |     0.0493us |     0.4228us |

A handful of things like Span.Reverse and Vector.Normalize are obviously improved, some are worse like Span.IndexOf, and some of these timing samples just make no sense and make me wonder if browser-bench is broken on my hardware.